+++ src/types/types.ts
@@ -2151,7 +2151,8 @@ export enum WebhookType {
   DROPPED_TRANSACTION = 'DROPPED_TRANSACTION',
   ADDRESS_ACTIVITY = 'ADDRESS_ACTIVITY',
   NFT_ACTIVITY = 'NFT_ACTIVITY',
-  NFT_METADATA_UPDATE = 'NFT_METADATA_UPDATE'
+  NFT_METADATA_UPDATE = 'NFT_METADATA_UPDATE',
+  GRAPHQL = 'GRAPHQL'
 }
 
 /**
@@ -2199,6 +2200,16 @@ export interface NftMetadataUpdateWebhook extends Webhook {
   type: WebhookType.NFT_METADATA_UPDATE;
 }
 
+/**
+ * The Custom Webhook can track any event on every block (think transfers, staking,
+ * minting, burning, approvals, etc.)
+ * This can be used to notify your app with real time changes whenever an
+ * EOA or a smart contract performs any action on-chain.
+ */
+export interface CustomWebhook extends Webhook {
+  type: WebhookType.GRAPHQL;
+}
+
 /** The response for a {@link NotifyNamespace.getAllWebhooks} method. */
 export interface GetAllWebhooksResponse {
   /** All webhooks attached to the provided auth token. */
@@ -2226,6 +2237,12 @@ export interface AddressActivityResponse {
   pageKey?: string;
 }
 
+/** Response object for the {@link NotifyNamespace.getGraphqlQuery} method. */
+export interface CustomWebhookConfig {
+  /** The graphql query for the webhook. */
+  graphqlQuery: string;
+}
+
 /**
  * Params to pass in when calling {@link NotifyNamespace.createWebhook} in order
  * to create a {@link MinedTransactionWebhook} or {@link DroppedTransactionWebhook}.
@@ -2249,6 +2266,20 @@ export interface TransactionWebhookParams {
 export interface NftWebhookParams {
   /** Array of NFT filters the webhook should track. */
   filters: NftFilter[];
+  /**
+   * Optional network to create the webhook on. If omitted, the webhook will be
+   * created on network of the app provided in the api key config.
+   */
+  network?: Network;
+}
+
+/**
+ * Params to pass in when calling {@link NotifyNamespace.createWebhook} in order
+ * to create a {@link CustomWebhook}
+ */
+export interface CustomWebhookParams {
+  /** GraphQL query */
+  graphqlQuery: string;
   /**
    * Optional network to create the webhook on. If omitted, the webhook will be
    * created on network of the app provided in the api key config.
@@ -2357,6 +2388,12 @@ export type NftWebhookUpdate =
 export type NftMetadataWebhookUpdate =
   | WebhookStatusUpdate
   | RequireAtLeastOne<WebhookNftMetadataFilterUpdate>;
+
+/**
+ * Params object when calling {@link NotifyNamespace.updateWebhook} to update a
+ * {@link CustomWebhook}.
+ */
+export type CustomWebhookUpdate = WebhookStatusUpdate;
 
 /**
  * Params object when calling {@link NotifyNamespace.updateWebhook} to update a